feat(http+gateway/C012): C012 P-D handler + WS payload + 集成测试 int64 → string
承接 P-C（commit ad2398b）。

http handler 层（23 文件）：
- pathInt64 / parsePathID 两个 helper 改返回 (string, bool)（保留函数名以减小 diff）
- 删除 strconv.ParseInt(c.Param("id"), 10, 64) 5 处（路径参数无需校验数字）
- presence batch endpoint 改 []int64 → []string，去掉 ParseInt 校验
- search.go ChannelID query param 直接 string 传透
- read_stats.go parseInt64CSV → parseStringCSV
- sendMessageReq / forwardMessageReq / saveAnnouncementReq / createApprovalReq /
  createTopicReq / sendUrgentReq / createScheduledReq / friendshipIDReq /
  syncChannelEntry / syncChannelResult / batch send inline struct：所有
  ChannelID/MessageID/ReplyTo/FileIDs/ID/RootMessageID/FriendshipID 字段
  int64 → string；零值校验从 == 0 改 == ""
- approval ListPending/ListMine cursor int64 → c.Query("cursor") 直接传 string
- notification ListReceived/ListSent cursor 同上
- scheduled List cursor 同上
- MessagePusher / ReadSyncPusher / ChannelEventPusher / MessageEventBroadcaster /
  ReactionEventPusher 5 个接口 channelID 形参 int64 → string

gateway 层：
- types.go：PushMsgPayload (ChannelID + ServerID) / SendPayload / SendACKPayload /
  SyncChannelState / ReadSyncPayload / ChannelEventPayload / ChannelClosedPayload /
  ChannelMemberUpdatedPayload / ChannelSchedulePayload (ChannelID + ScheduledID)
  全部 ID 字段 int64 → string；Seq 保留 int64
- ChannelSeqStore interface 返回 map[string]int64
- conn.go: knownSeq map[int64]int64 → map[string]int64；UpdateKnownSeq/
  KnownSeqFor/KnownSeqs 形参/返回类型同步
- heartbeat.go: 删 fmt.Sprintf("%d", chID) 转换，chID 已是 string
- ws_handler.go: WsMemberLister.ListMembers/GetMember 形参 channelID string，
  SendPayload.ChannelID 零值校验 == ""，trace span attribute.Int64 → String 3 处
  (channel_id / server_msg_id)；PushID format %d-%d → %s-%d
- ping handler 删 Sscanf 数字解析，ChannelSeqs key 已是 string

repo 层补丁（unblock 测试）：
- 9 个 primaryKey TEXT 字段加 default:gen_random_uuid()::text gorm tag，
  让 GORM 在 INSERT 时跳过空 ID，让 DB DEFAULT 生效（C012 P-A 已 set DEFAULT，
  但 GORM 不识别 model 上的 default 标签会发空字符串覆盖 DB default）
- 8 个 .First(&m, id) → .First(&m, "id = ?", id)，避免 GORM 把 string ID 当
  raw SQL 表达式注入

cmd 层（最小必要修复 hub adapter，仍属于 P-E 边界）：
- cmd/gateway/main.go: 6 个 hub adapter (hubMessagePusher / hubReadSyncer /
  hubChannelEventPusher / hubEventBroadcaster BroadcastMemberEvent /
  hubEventBroadcaster BroadcastToMembers / hubReactionPusher) channelID 形参
  int64 → string；删 strconv.FormatInt(channelID, 10) → 直接传 channelID
- cmd/message/main.go: incomingMessage / deliveryEvent struct ChannelID/
  ReplyTo/ServerMsgID int64 → string；partitionKey 直接用 msg.ChannelID；
  PushID fmt %d → %s
- 两个 main 的 strconv import 去掉

测试 fixture：
- internal/testutil 隐式跟随 repo / service 改造（C012 P-B/P-C 已落）
- tests/integration/m4_harness_test.go: seedDM/seedGroup 返回 string；
  seedMessage 形参 channelID string；6 个 local adapter (BroadcastToMembers /
  PushChannelEvent / BroadcastReaction / BroadcastMessage / PushReadSync)
  channelID 形参 int64 → string
- 21 个 m4_*_test.go：
  * 103 处 strconv.FormatInt(<id>, 10) → <id>
  * 34 处 int64(x.Value("id").Number().Raw()) → x.Value("id").String().Raw()
  * 23 处 .Number().IsEqual(float64(<id>)) → .String().IsEqual(<id>)（仅 ID
    字段；status / seq / role 等 numeric 字段保留 Number().IsEqual(float64())）
  * 16 处 float64(<id>) 裸 cast 去掉（require.Equal 路径下）
  * 10 处 []int64{ChannelID/MessageID} → []string{...}
  * pathInt64s helper 改 (v string) string 身份函数
  * 部分 channelID / annID / approvalEnv.channelID / fixture struct 类型
    int64 → string
  * 多处 strconv import 删除

测试：
- go build ./...                       PASS  ✅
- go vet ./...                         PASS  ✅
- go test -tags=integration -count=1 -v: 195 RUN / 162 PASS / 30 FAIL / 3 SKIP
  - 18 个 fail 是 C5_BadRequest / C5_InvalidPath 测试，明确测「非数字 path → 400」
    的旧 int64 校验，C012 后 path param 接受任意 string ——属于 C012 §3.4 预期
    的测试语义变化，留给 C014 测试覆盖收尾时统一调整（C014 spec 已声明）
  - 12 个 fail 与 C012 ID 类型无关（auth team_id header / WS push 回放时序 /
    notification quickreply / scheduled），上溯到 v0.7.4 鉴权 / 推送回路改造的
    pre-existing test design 漂移，留给 P-E + C014

边界：
- 用户 WIP（CORS / IM_REDIS_CLUSTER / sync.go 中文标点 prompt injection /
  message.go gofmt / docker-compose pulsar-init / CLAUDE.md C009-C011 / Makefile
  IM_REDIS_CLUSTER override）在 P-D 期间 stash 保留，未提交
- 不改 migration / 其它 cmd 文件 / docs / .idea
- Seq / FileSize / cursor 业务计数仍 int64
- 未跑 mockery（service 接口未改）

引用 spec：docs/harness/C012-id-type-string-migration.md §3.2 + §3.5 P-D

下一步 P-E（cmd/sync + cmd/v4-client + Makefile + 修剩余 30 集成测试 + mockery）。

+++ server/tests/integration/m4_ws_channel_friend_events_test.go
@@ -112,7 +112,7 @@ func TestM4WSChannelInfoUpdated_HappyPath(t *testing.T) {
 	frame := wcA.expectFrame(gateway.TypeChannelInfoUpdated, 5*time.Second)
 	var snap map[string]any
 	decodePayload(t, frame, &snap)
-	require.Equal(t, float64(chID), snap["id"], "channel_info_updated id must match")
+	require.Equal(t, chID, snap["id"], "channel_info_updated id must match")
 	require.Equal(t, "after-patch", snap["notice"], "channel_info_updated must carry new notice")
 }
